refactor : Parser 유틸 클래스 작성 - parseFilterCondition 메소드 이동
